@@ -8,7 +8,7 @@ const appRoutes: Routes = [
 
 
 @NgModule({
-  imports: [RouterModule.forRoot(appRoutes)],
+  imports: [RouterModule.forRoot(appRoutes, {})],
   exports: [RouterModule]
 })
 export class AppRoutingModule { }
